packaging: Disable HTTP and HTTPS when using Apache as proxy
This patch changes the setup application and the JBoss
configuration file so that the HTTP and HTTPS connectors will be
completely disabled and only the AJP connector will be enabled if
Apache is used as a proxy.

On the other hand, if Apache is not used as a proxy the HTTP and
HTTPS connectors will be enabled but the AJP connector will be
disabled.

This reduces the use of resources in the application server
(thread pools) and reduces also the attack surface.

It is convenient to be able to enable/disable the connectors at
will after the installation. This can be done changing the
/etc/sysconfig/ovirt-engine file, without running the setup again.
For example, in order to enable the HTTP connector it is enough to
add (or just uncomment) the following lines:

  ENGINE_HTTP_ENABLED=true
  ENGINE_HTTP_PORT=8080

Then restart the engine and the connector will be available. Same
for the HTTPS and AJP connectors.

Change-Id: I0d9443d18f1eee928a2fe594527ac0d59c14a5df
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>