glance.pp overrides sql_connection for glance-api.conf also
Previously, the sql_connection config was overridden for the
glance-registry.conf only, where in the glance-api.conf this
is left to default to a non-existent sqlite DB.

This causes an issue with the v2 API is used - in this case,
the glance-api service accesses the DB directly instead of
delegating to the registry service. Since the sqlite DB is
non-existent, such v2 API calls fail with 500 ServerError.

This is especially problematic with cinder, as it tries the
glance v2 API by default when creating a bootable volume from
an image, see:

  https://bugzilla.redhat.com/888241