Adding homedir custom fact
The location of the keystonerc_admin file was hard coded to /root,
this sometimes is not the home directory of root. We need a custom
facter fact to get the correct home directory.

This commit
o adds a facts directory to contain facter facts
o adds a home_dir fact
o uses this new fact for the location of the file keystonerc_admin

Change-Id: Ifb7003072939d4c95f8ee65b721bfe31311d298c