Update to new version of puppetlab-swift module
XFS storage devices were working
https://bugzilla.redhat.com/show_bug.cgi?id=906471

New swift module has a cherrry picked commit
https://github.com/packstack/puppetlabs-swift/commits/packstack/folsom

== From cherry picked commit ==
fix bug setting device
The device set for the storage mount point was
hard-coded to /dev/$name. This value is only
valid when no device is supplied to the storage
xfs setup.

This commit updates it to use the target_device
which is either the device that was passed, or
/dev/$name if not device is specified.

Change-Id: Id5a637ca7dbf8f4968d3fc39b420f078d51fbe3c