Adding a install option to install in one step
This commit adds an option to facilitate people trying to install openstack in
a single step without first generating an answerfile
Users can run the command
$ packstack --install-hosts=192.168.0.22,192.168.0.31

If a singe host is supplied then an all in one setup will be installed, if multiple
hosts are provided, then the first will be a controler and the others will become
nova compute nodes. For all of the other value defaults are used unless supplied in the
command line e.g.
$ packstack --install-hosts=192.168.0.22,192.168.0.31 --novacompute-privif=eth0 \
  --novanetwork-pubif=eth0 --novanetwork-privif=eth0

An answerfile is also generated, this should be used if packstack needs to be run again

https://bugzilla.redhat.com/show_bug.cgi?id=892247

Change-Id: Idd97183177dc349396274f427d068a30ed59faae