Set verbose=True by default in quantum.conf
This updates the quantum submodule because of a recent change that
makes it preferable to set boolean values via native puppet
true/false instead of via strings like 'True'/'False'.

As a side effect, this update also fixes the use of log_file in
quantum.conf instead of log_dir.

Change-Id: I3b0a7662e9840fb580bcdfcd72dda8bd234d6804