Add global option --exclude-servers=EXCLUDE_SERVERS
This is a comma separated list of servers to be excluded from
installation in case you are running Packstack the second time
with the same answer file and don't want Packstack to touch
these servers.

For example, suppose that you've already installed the following
hosts:

  - Controller node : 10.0.0.99
  - Compute nodes   : 10.0.0.100 - 10.0.0.110

and you want add one more compute node 10.0.0.111.

Then you can reuse the previous answer file with the following
modification:

 - Add 10.0.0.111 to NOVACOMPUTE_HOSTS
 - Add existing hosts (10.0.0.99 - 10.0.0.110) to EXCLUDE_SERVERS

This avoids applying puppet modules to existing hosts again.

(In theory, because of idempotentness of puppet manifests,
applying the same modules does no harm. But it is a waste of time
and has a potential risk of breaking manual configurations which
has been done after the packstack installation.)

Change-Id: Ib3dcd94a25a286f82cc0fcedebd1824b1d27d864