Fix provider network option
Setting external_network_bridge = false in the manifest resulted
in the value 'False' being set in l3_agent.ini. Now an empty string
is passed in. It should be noted that the puppet module does not
handle empty values in config files well and that successive runs
will ignore the fact that the setting is there at all, resulting
in multiple external_network_bridge settings being added to the
config file. This is something that will need to be addressed
separately.

In addition, this patch ensures that the CONFIG_QUANTUM_OVS_IFACES
are applied everywhere the OVS agent gets installed.

Change-Id: I9481a9d94300d3aa6d713f39c1f6d7430d4af900