Remove libvirt's default network (i.e. virbr0) to avoid confusion
When libvirt is installed, a default network is created. The user
will notice a new interface, virbr0 with the IPv4 192.168.122.1.
Now when configuring things and most particularly when debugging
(e.g. complex Neutron-based) networking issues, this extra interface
and IP (range) tends to confuse people. Therefore, Packstack should
better remove it as it's completely unecessary on a OpenStack
compute node.

Change-Id: Icfa81a9124ec4cf623528219edf5d73ca85cda72