Default to use emX instead of ethX on Fedora.
Fedora by default uses network interface names based on their
physical location instead of the legacy ethX naming. While various
names for the different physical options exist, Fedora is mostly
installed on commodity hardware (or VMs). Therefore, it is best
assumed that the devices are embedded interfaces and therefore
called emX on this platform. It should further be noted, that the
interface numbering starts at 1 instead of 0 with this convention.

Change-Id: I8035715eea61b7bf0b8413bbded5d95b1eb4e1a8