Adds Firewall rules for mysql access.
Allows only hosts that need to connect to mysql.

This patch is for mysql only, rules for the other
services are being developed since the bug
addresses the entire openstack installation.

Fixed and apply review comments

Changed the firewall template so it can be used
by the other plugins.

Fixes (partially): rhbz#1002063

Change-Id: I6639ebd5438d47c92d8cabb91f94e72b61426b98