Adds Firewall rules for swift access.
Allows connection to swift and rsync from
this hosts: CONFIG_SWIFT_STORAGE_HOSTS and
CONFIG_SWIFT_PROXY_HOSTS

This patch is for swift only, rules for the other
services are being developed since the bug
addresses the entire openstack installation.

The swift proxy service was not restricted
because it's the connection to the outside world.

Fixes (partially): rhbz#1002063

Change-Id: I1f8e46ff4eebe85d2d0a4bcded8e6ad8ed664a9b