Ensure horizon SSL is configured when enabled.
Ensures the creation of
/etc/httpd/conf.d/ssl.conf
with the minimum requiremets for the SSL service
to work.

Note: The real problem is that the puppet module
that installs apache's mod_ssl does not install
the ssl.conf file (the manual package installation
does).
The module code shows no hint of a problem so i
have to contact the puppet module maintainer to
help me. Since this can take a little time and the
SSL option is important i wrote this patch to
address the issue.

Change-Id: I886b027b3e63901dc288f1fec05b3aa3ce91b4dd
Fixes: rhbz#966094