Don't override cinder's default service type.
Patch I90eab888f727a0cb5b5b5e00efc593000df98c55 introduced the creation
of a service type and endpoint for cinder's API v2. Although using v2 is
fine, the patch introduced it by overriding the default service type for
cinder.

This is an issue because cinder client is still using v1 as the default
version. This patch fixes that issue.

Change-Id: Ie67798e4a1cb5a18dbd630415d13ed6c1a60e89b