Explicitly include the firewall module
Puppet automatically loads custom resources types provided by modules,
but does not automatically run the manifests provided by the module.
Module manifests are only activated by explicitly "including" the
module. The "firewall" module manifests take care of ensuring that the
necessary packages are installed to support the commands used in the
firewall resource type.

This patch includes the firewall module in the prescript.pp template,
which ensures that the appropriate packages are loaded before firewall
resources are realized.

This patch is related to the following upstream pull request for the
firewall module, which ensures that the proper packages are installed on
Fedora and RHEL 7:

https://github.com/puppetlabs/puppetlabs-firewall/pull/338

Change-Id: I13ba5ae20b3c9d4a605b1358bd343aa730c63eaf
Closes-bug: 1297857