don't install l2 agent on api node by default
The q_host is the sum of api host, compute host, and network host. And
it is not necessary to install l2 agent on api node by default, because
api node is only responsible to api request routing. This patch removes
the default behavior of install l2 agent on api node.

Change-Id: I89c3a1be9c5f336c2eceb3773c3bcff2d9c1c61b