Remove ServerAlias definitions
Parameter fqdn of class horizon is used both for ALLOWED_HOSTS in settings_local.py
and for ServerAlias directives in vhost.conf which is breaking server accessibility.
We need ALLOWED_HOSTS values, but we have to avoid ServerAlias definitions.
For now we will use this wildcard hack until puppet-horizon will have separate
parameter for each config.

Change-Id: I77f1baa4f9c0d94b00294f17978d8383a2386d7a
Fixes: rhbz#1119920