Use a valid e-mail for admin user in keystone
The default installation uses test@test.com as the e-mail address
for the admin user created in keystone. This patch changes that
e-mail to root@localhost, which is consistent with the addresses
set for the rest of the users.

Fixes bz#1020199

Change-Id: I480f1bd4188d65b5c32de113a07cbdd5ab9beb9d
(cherry picked from commit 264250b626d71ef6e2eaa4a31c57fad037d49ffb)