Generate answer file only when needed
Change Ic5947567599c8b221b7a9e60acb4708429507741 introduced some
modifications on the way the answer file was generated. This
results in an answer file being generated even if the user
specified --answer-file as a parameter. This patch fixes that by
not generating an answer file when --answer-file is specified.

Change-Id: I52e59d45a4b63a2f1ed70a154bca6b705ed435d2