Fix Neutron FWaaS configuration
The Neutron FWaaS driver was always loaded, even if
CONFIG_NEUTRON_FWAAS was set to "n". This, and the fact that it
was configured after the L3 agent was started, caused
neutron-l3-agent to be down after Packstack execution. This patch
fixes it by setting the right configuration order and making sure
the driver is only configured if needed.

Change-Id: Idc37f53d64ddf5e08e3229be9241bae2f8ac9e67