Always enable EPEL repo when installing RDO
The EPEL repository is a requirement for RDO installation. This
requirement was previously handled by a downstream patch on the
package changing the default value of CONFIG_USE_EPEL, but this is
complex to maintain. This patch overrides the specified value of
CONFIG_USE_EPEL when RDO is being installed, while respecting it
for other distributions.

Change-Id: I393128333abacc841bf32adcf104d4635f616539