Allow alternate locations for Cirros image
When creating the demo environment, the Cirros image was downloaded
from a hardcoded location. This can create issues in environments
placed behind a proxy or isolated.

This patch adds a new config option (CONFIG_PROVISION_DEMO_CIRROS_URL)
where a local file path or remote URL can be specified. Be aware that
it requires a fix to puppet-glance to work
(https://review.openstack.org/132813).

Change-Id: I2e6a3ba9822624d16c63f88c30d24d26d6d0c3e0
Fixes: bz#1147716