redis-based group membership coordination for ceilometer
If CONFIG_CEILOMETER_COORDINATION_BACKEND is set to 'redis'
then a redis manifest is created to install redis on the
appropriate host and configure the ceilometer-central-agent
and ceilometer-alarm-evaluator to use it for group membership
coordination.

The other valid option is 'none', in which case no coordination
will be done. 'redis' is the default.

Change-Id: I07261f5fd433c13a57a614c15b13f0cad6c79d6a