Add support for redis sentinel and slaves
If the correct configuration settings are set in the answer file
redis and redis-sentinel will be installed and configured on slave
and sentinel hosts. The sentinels will monitor the master and
keep track of the slaves. If the master dies, then the sentinels
will consensually establish a new master.

Change-Id: I7acc5e4f7bb99692bfe1c8694e8ddb0c472160f4
(cherry picked from commit 3fd2f89875510098595642c171e14ba2e51f727f)