Update redis-sentinel support to handle multiple sentinels
If there are multiple sentinels we would like to configure the
coordination backend to use all of them. This change makes that
possible by constructing a coordination url that lists all the
sentinel hosts: one in the netloc portion of the url and the rest
as sentinel_fallback parameters. If no sentinel hosts are provided,
none are used. If only one is provided, only the first host is set,
no fallbacks.

This change also ensures that the configured sentinel port is respected
in the generated sentinel configurations and in the ceilometer
coordination url and that redis slaves bind to an appropriate host.

This URL format will be present in the forthcoming tooz 0.11.

Change-Id: I49edd2143b800ee1c3020fc04ed5ad817ddffe6c