Fixes heat missing configuration file
If the keystone_heat manifest is added to the keystone.pp
manifest it will fail because of missing heat.conf file.
This is due the fact that heat is not installed yet,
this manifest should be appended to the heat.pp manifest to
avoid this error.

For this to work, this puppet-heat patch is needed
https://review.openstack.org/146734

Change-Id: I765cb7fa7bd8d346235fa2c1fc3ca6eea7b9f636