Only use sentinel when redis ha is chosen
When running in allinone mode, packstack will set any config
variable ending with _HOST to have a value of the controller host.
This was causing the ceilometer coordination_url to inadvertently be
set to use sentinel even when redis high availability was not configured.

By adjusting the template to be more explicit, this is no longer the
case.

Change-Id: I281e9de507f68f4f2a5941b14b5696c91c12fdbb
Closes-Bug: #14336643