Merge serverprep and prescript plugins
After commit 012a812609b32e3bb0ad6fc69d53d0486f8691fa was merged,
some configurations could be broken, such as those requesting EPEL
repos to be configured by Packstack, as Puppet was installed before
the repo was enabled. The best way to fix this is by merging the
serverprep and prescript plugins into one.

Together with the required adjustments for the merge, a new
auxiliary function is created to detect the OS name and release, to
allow some functions previously in the serverprep plugin to run
before the host facts are collected.

This patch is rather large, so extensive testing should be applied
before merging. You have been warned :).

Change-Id: I0b89349827c403070b6daddbbfd0dab535a95c05