Add processor process_string_nofloat
In some corner cases, the Heat auth_encryption_key value set by
the user or randomly can be convertible into a float number. This
will make the puppet-heat module to complain, because the size()
function from puppetlabs-stdlib will check that before returning
the array size.

Fixing this by implementing a new processor, that will check for
those cases and change the value to a random 16 char hex string if
needed.

Change-Id: I6d9541692f9f87b100ae1e4651038297c802cbf4