Fix br-ext IP configuration
The provision plugin adds the_br_ext suffix to the neutron
bridge configuration which results in a inicialization problem.

There is no need for the configuration script since openvswitch
manages the creation of the bridge interface.

Setting CONFIG_NEUTRON_OVS_BRIDGE_IFACES to a suitable value will
trigger the creation of the interface (defaulted to br-ex:eth0).

Also, for the proper port creation CONFIG_NEUTRON_OVS_BRIDGE_MAPPINGS
must be set to the proper value.

Change-Id: Ie6c2ed966627f6b9f3c485af51270176a867a527
Closes-Bug: 1459413