Fix Sahara support
After https://review.openstack.org/192721 was merged, some small
adjustments are needed to enable Sahara support in Packstack. If
not, only openstack-common will be installed and the service will
be configured, but no services will be started.

This commit adds sahara-api and sahara-engine, so Sahara will use
the suggested split mode.

Fixes: rhbz#1265661

Change-Id: Ib403b5bcf08c5ac71f073992cffbe3c634818c61