[vmware] Add support for multiple VMware vCenetr clusters
Previously packstack didn't support multiple vCenter clusters. The
option CONFIG_VCENTER_CLUSTER_NAMES now ensures that if one cluster
is specified it gets mapped to all computes and if multiple clusters
are specified it'll map one cluster to one compute. This means that if
multiple clusters are specified it is required to have same number of
nova computes.

Change-Id: Ide73f21fca5e6f51bd43aafe4a2b8554b92b1a87
Resolves: rhbz#1110891