Provide framework for integration testing with run_tests.sh
The goal is to get run_tests.sh in a shape where it can be used and
implemented in a jenkins gate job to eventually get packstack to
gate against itself.

The scenarios provided in the test matrix aligns with the ones
that are integration tested in puppet-openstack for the sake of
standardization.

Change-Id: I68673a1f7e0584916a814296c9184b927a30aab6