Add --service-workers parameter and scale mariadb connections
This is a new config that will define the value to be used
in parameters passed to different services as an amount of
workers or threads to use.
Most puppet modules default the amount of workers or threads
to $::processorcount and this yields fairly high RAM usage on
machines with a lot of cores. Since Packstack did not set this,
we would use the default provided by the module.

8 cores with fairly minimal services yields >8GB RAM usage after
tempest tests.
If we set --service-workers to 2, we peak at around 6.5GB RAM for
the same Packstack parameters.

We will still default to $::processorcount for backwards
compatibility but this new parameter can be used to lower the
RAM usage where there are constraints (i.e, upstream gate)

Additionally, instead of hardcoding a max_connections of 1024
which provides a ceiling of ~3GB RAM, we now instead default to
a sane value depending on the amount of service workers (n*128).
This will still yield the same 1024 max_connections for 8 cores
but would set a max_connection of 256 when using 2 service workers
for example.

Change-Id: I2153ce522f227678d5318080dfb4ef171b36cbf7