Fix bridge creation for compute nodes when using VLAN
Commit 51fb0c52003745739700afd94ac8725c978dec41 changed the OVS
bridge creation to only happen in network nodes. This breaks setups
with VLAN tenant networks, where a bridge is required in compute
nodes.

This patch fixes this, and also creates bridges when using VLAN
networking. Depending on the configuration, it may create an unneeded
bridge in compute nodes for the external network, if both are
configured, e.g. br-ex for external traffic and br-vlan for tenant
networking. This should be handled in a separate patch, since there
is currently no way to distinguish which bridge is needed in compute
nodes and which one is not.

This fixes bz#1325637

Change-Id: Ia6e982943db0af5465f9e18261e2803a22a0ace0