Simplify apache::listen for services relying on httpd
Due to the way Packstack runs Puppet manifests, we are regularly
having issues related to usage of apache::listen. Using it means
the existing configuration in ports.conf is removed with every new
Puppet run, so we need to maintain an ugly hack in services relying
on httpd to ensure no port is lost.

To simplify this, we are consolidating all apache::listen calls in
one template (apache_ports.pp), which is included by any service
requiring Apache. Also, a new plugin (apache_060.py), will make sure
Apache is installed with a default configuration, so any service
using it later finds the same starting point.

While it does not fix the design issue, it provides a single place to
specify listen ports, and will avoid further issues.

Change-Id: Ic8ed069f4991602d125f40eaad8f050284d680e1