Create only required OVS bridges in compute nodes
Packstack currently configures the same set of OVS switches in network and compute
nodes. In some topologies neutron requires different OVS bridges in different nodes.
as br-ex only in network nodes and br-provider in all nodes for provider networks
or br-vlan for vlan-based tenant networks.

This patch add an additional parameter to packstack:

- CONFIG_NEUTRON_OVS_BRIDGE_COMPUTE

Which must be used to specify the list of bridges that must be created in the
compute nodes. Interfaces will be connected to these switches as specified in
CONFIG_NEUTRON_OVS_BRIDGE_IFACES parameter and mappings added as declared in
CONFIG_NEUTRON_OVS_BRIDGE_MAPPINGS.

Closes-Bug: 1569360

Change-Id: I647ffc1e9bd2ced1be994997e097e3b641964b64