Add support for Packstack to manage a default set of flavors
Nova no longer provides flavors by default in Newton.
This commit adds a parameter to have Packstack manage a
default set of flavors and it defaults to true to keep
backwards compatibility.

This implementation will need to be revisited but for now this
should allow us to progress in Newton.

Change-Id: I4f1bdff77cc3ea90997fe69a80ed314838a5f075