Change way of cloning OPM from source
Up to now, we were using a consolidated openstack-puppet-modules
repository. Since we are moving to individual packages, that repo
will no longer be maintained, and we will switch to cloning the
puppet modules themselves.

The code to install Puppet modules has been adapted from
https://github.com/openstack/puppet-openstack-integration/blob/master/functions,
to also benefit from the Zuul cache for OpenStack puppet modules.

In the current patch, all Puppet modules are taken from their
master branches. This is consistent with the way they are handled
in RDO Trunk packaging.

Change-Id: I7cc6f5ec138dd7e8d9ef06fc7e59ad8e3d5d58e8