Remove OPM dependency resolution in prescript
Previously, the dependencies for the openstack-puppet-modules and
openstack-packstack-puppet RPMs were being manually resolved and
installed in any non-installer node (usually the controller). Since
Newton, the OPM package is a meta-package, that includes all the
puppet-* packages as dependencies. In this context, Packstack would
try to install all puppet-* packages in all nodes, which is not
needed and causes issues.

We are solving this by removing the dependency resolution code, adding
a new array to include OPM dependencies (only rubygem-json as of now)
and installing them when needed. Note that, if a dependency is added
to any puppet package, it will have to be added to Packstack as well.
CI will notify us, though.

Change-Id: I54d0c0296c0b57946e7af1f7fd13dae35c1c1a36