add wsgi threads for gnocchi::api
Currently the number of gnocchi api threads running equals the
number of CPUs in the system. This leads to wasted resources.

This patch makes that number of threads configurable via
CONFIG_SERVICE_WORKERS. This is the variable that controls the number
of workers in other packages. In gnocchi::api it controls the number
of threads.

Change-Id: Ic332c06fa8e304226e113d5f17a34b25290640e3