Use pre-cached images for tests (or pre-cache prior to testing)
Cirros downloads have been a big problem of unreliability in testing
Packstack in different environments.

This aims to use pre-cached Cirros image files (in OpenStack or RDO
gate) or it will download the images with wget (with retries) prior
to running the tests.

This will prove faster if there are pre-cached images available
and more reliable in case the images need to be downloaded.

The change is backwards compatible and affects testing from
run_tests.sh only. The glance_image provider from puppet-glance
accepts files or URLs, detecting if it's a path by leading slashes
in from the location.
The *_URL parameters are arguably no longer accurate but this can
be fixed in a future patch if we see the need to.

Change-Id: Ibd19fb69e080957165594953687efe5f096f93b2