Use ADDITIONAL_ARGS to pass extra commands to Packstack scenarios
This is a generic variable that'll allow to pass extra commands to
Packstack integration testing scenarios so that we do not have to
parameterize everything and extend testing easily if required.

For example, this will allow testing frameworks to send the extra
argument "--enable-rdo-testing=y" in order to test Packstack against
the testing repositories.

Change-Id: I5b510aabe6749e329b19ef178e6b8870ee66c1a7