Remove unix socket from redis configuration
Since redis-3.2.4-1.el7, /var/run/redis is not longer
created, so redis fails to start with default configuration
in puppet-redis because it can't create socket file.

This patch removes unix socket from redis configuration
as it's the default in redis.conf file and it's not
required for openstack services (they connect via TCP
socket).

Change-Id: Ifd030b019becb1db6c98a09fe7fad0493e685516
Closes-Bug: 1633498