Support puppetlabs hiera.yaml update
Previously the code to update the hiera data location for packstack
assumed /etc/puppet/hiera.yaml which is no longer used in newer versions
of puppet provided by puppetlabs. This change updates the prescript with
two things. The first it skips creating the symlink to /etc/hiera.yaml
if /etc/hiera.yaml does not exist. And second it uses the puppet config
command to get the hiera.yaml location to update.

Change-Id: Icd931f36fc14fefbcd43a266f0cd538074723adf