Switch to fernet tokens by default
In the upcoming puppet keystone release, the default is being switched
to fernet. This change updates packstack to properly handle the fernet
provider configuration for keystone and makes it the new default.
Since the upstream will be enabling the fernet token setup by default,
packstack needs to properly turn off the fernet setup when using uuid.

Change-Id: Id9f6d54e7e5190c3a01f60c511d8a470436bff7d