Avoid error if logFile is not defined
If there is an error initializing a plugin or checking the command-line
arguments, we may use the logFile variable without having
initialized it, which causes an exception.

In those cases there will be no log file, so let's catch the
exception and ignore it if it happens.

Change-Id: Icaeb2ae780da9802591e9a063a71dd339b396b93
Closes-Bug: #1635770
(cherry picked from commit 6f6b5aa90999db43c2b364ac20c456387ef44eea)