Add update step after enabling OpenStack repos
In specific cases, there may be conflicts between packages
provided by Centos and RDO repos. In that case RDO repos
should be used.

To avoid any of these issues, this patch updates packages
right after enabling the repos.

Note this will be required when gate jobs move to CentOS 7.3
because of conflicts between mariadb-libs in CentOS and
mariadb-common in RDO repos.

Change-Id: Icaeed8602378780f2c9bc8b22621c09b65bf95a8