Add options to display parameter specific messages
This patch introduces MESSAGE and MESSSAGE_VALUES to
parameters. When a user selects a value included in
MESSAGE_VALUES for a parameter with MESSAGE option,
this message will be shown to users at the end
of packstack execution.

Example use case is showing a note alerting that a service
will be removed in next release if user enables it.

Change-Id: I1331de921e1f5f1aecf991cb83fa1509f29b3ee7