Align the stars to fix CI
This is a combination of two fixes:

1- Update version of hacking used in tests

We were using an old version of hacking that did not support
pbr==2.0.0, and this is creating issues running flake8.

Since we are now using a newer version, a couple updates are required
to get the tests to pass.

2- Remove the workaround for the Nova placement WSGI file

https://review.openstack.org/440638 added a workaround on the Puppet
side, so it was conflicting with ours.

Change-Id: I7384a7b5612866b2a70643ae4203a1848fdf8679