Removed test_plugin_prescript.py
Packstack gerrit gating does not check unit tests on RHEL.
Because of that the only test in `test_plugin_prescript.py`
is not failing. The test name is `test_rhn_creds_quoted`.
On non-RHEL systems test is simply skipped.

Currently the test does not make any sense to exist, hence
the removal.

Fixes-Bug: #1694469
Change-Id: I0e65cd76f5005edf0074cd378b2314298a0670db
Signed-off-by: Maciej Kucia <maciej@kucia.net>