Use qcow2 as disk format for ami images
According to [1], 'ami' is a valid disk format for Glance images.
However, after [2] we are failing to boot from volume on AMI
images because qemu-img convert does not recognize it as a valid
disk format. The image is really qcow2, so let's use it as a
workaround (see [3] for the bug description).

[1] - https://docs.openstack.org/developer/glance/formats.html
[2] - https://review.openstack.org/453341
[3] - https://bugs.launchpad.net/cinder/+bug/1693926

Change-Id: I2dbf4736ebe4cda776fc51de62af462b800f918b