Define key_manager/api_class to a cass full path
https://review.openstack.org/#/c/505752 has switched to use
key_manager.backend which is set by default to 'barbican'.
However, until fix in cinder [1] is merged, cinder expects it
to be a full class path. Castellan uses key_manager/api_class
to resolve key_manager.backend so, setting this parameter to
an existing class will fix jobs until [1] is merged.

A similar workaround has been pushed to p-o-i in [2]

[1] https://review.openstack.org/#/c/468252/
[2] https://review.openstack.org/#/c/507567/

Change-Id: I71755a9f5fb29116167c552988284903a850dfd8