Use system arch to determine what cirros images to use in run_tests.sh
Currently run_tests.sh hardcodes what images are downloaded and used
for testing. This patch grabs the system's arch and uses it to build
the urls and filenames for the cirros images. Additionally, ppc64le
images are not provided until 4.0 so switching cirros version based
on arch is also added here.

Change-Id: I73b65c566153c027988b5781870e44c47c6e8b1f