Fix Hiera keys with a dot
The new Puppet version does not recognize dots in hiera key names,
so we need to replace them with underscores.

Change-Id: I6642d5963a598151dedbbbd957449bb317ceaa59

Replace deprecated 'auth_uri' with 'www_authenticate_uri'

Deprecated 'auth_uri' parameter is being removed from
puppet modules in [1].
In Swift 'auth_uri' is still not deprecated, it's being
deprecated in [2], so that can be taken care later.
For Sahara use ::sahara::keystone::authtoken class to configure
keystone parameters as the one specified in '::sahara' are deprecated
for removal.

[1] https://review.openstack.org/#/q/topic:remove-auth-uri
[2] https://review.openstack.org/#/c/621138/

Co-Authored-By: Javier Pena <jpena@redhat.com>
Change-Id: I68edd9225b44938dc6c275bf464d0e328e814848