Pin sphinx to <2.0.0
The 2.0.0 release introduces a requirement for using
python3.5 but we are running this in python 2.7

Also pining puppet-rabbitmq as commit [1] has broken compatibility with
puppet < 5 because of the usage of data in modules. I'm pining to the
same commit used in RDO for Rocky.

[1] https://github.com/voxpupuli/puppet-rabbitmq/commit/cc78a33d98defdf0ef796d98ceb35b811168db5e

Change-Id: I055ab99e48e018d041a2aef7c8998f72ebd4ff5e
(cherry picked from commit d2b1c67d57035416bf04f04dabb4e927216836ff)