Set unixsocket to undef in redis class
After [1], redis class does not longer accepts boolean as value for
unixsocket, so let's undefine it. Note that this has the effect of
enabling unix socket in redis but I don't see a way to get it disabled
after the patch.

To minimize risk, i'm setting permissions of socket to 0700.

[1] https://github.com/voxpupuli/puppet-redis/commit/502ff38c8c035fa143e32dd4ee61b3ec68db461c

Change-Id: Ic967d9446ef68a0257b9448bc8f3f3b664d776e0