Use rabbitmq parameters for tcp_listen_options
Currently we are passing tcp_listen_options as part of config_variables.
However, puppet-rabbitmq creates this directive from specific options [1].
This is leading to configuring it twice with different options. This is
not valid for rabbitmq 3.8 which does not start if it detects it
configured twice.

[1] https://github.com/voxpupuli/puppet-rabbitmq/blob/master/templates/rabbitmq.config.erb#L34-L55

Change-Id: I22048b580fef254090c77f8ded58513b44e93c36
(cherry picked from commit 96bd0f7ca64c95d96d64b55b5f6c443e4d24343e)