Fix manage_rdo function for CentOS8
In CentOS 8 the rdo-release package uses %dist in the name, which is
currently breaking the regular expresion used to parse it.

Also, in CentOS 8, yum-config-manager is not providing the full repo
configuration by stdout when enabling a repo, so instead of parsing it,
let's just trust on standard return value from command which works as
expected. For CentOS 7 we need to keep the same behavior.

This patch is fixing it to work with both centos7 and 8

Change-Id: Iafd6d6373d608c1b0fb797895a9969da28049fec