Let apache module purge config
Otherwise can cause issues when enabling prefork as mpm_event is enabled
by default in httpd. A recent change in puppet-apache [1] has added
httpd.modules.conf directory to load, which includes default mpm config
that conflicts with prefork.

In the past we disabled purge because of issues with nagios [2], however
nagios deployment is not longer supported by packstack, so we can safely
enable apache config purge which is the default option.

[1] https://github.com/puppetlabs/puppetlabs-apache/commit/6425cc46a841fa6ea5bb3f34b8f31d86c14e6914
[2] https://github.com/redhat-openstack/packstack/commit/686625e17b19c31b9640c64263d9d3bb2bab48c7

Change-Id: Ifdccd8609658b3170b8356bddc5062f252677d40