Use netifaces instead of facter to find ip address
Currently, neutron_350.py is using parsed output of `facter -p` in
prescript_000.py which is actually not working fine in terms of getting
the ip of a interface.

This patch is using netifaces python library instead to get the ip
address which is much easier to use and reliable.

Note that netifaces was already a requirement in packstack so it's
included in requirements.txt and as package requirement.

Change-Id: Id4d2b4ca814129c485f491ed91bb6bb99e462114
Resolves: rhbz#1734354