Move to cirros version 0.5.1 by default
We are still using old 0.3.5 version, let's move to the most recent one.

This patch also removes usage of UEC cirros image for tempesa. This was
introduced long time ago [1] because of hitting issue reported in [2].
This should be fixed in newer cirros versions so let's make it simple
and do not use extra image only for tempest.

[1] https://review.opendev.org/#/c/322139/
[2] https://bugs.launchpad.net/cirros/+bug/1312199

Change-Id: I1bc286f6264538f4be920e5bf2b5d113c0b57280