Use magnum::keystone::keystone_auth for magnum config
Puppet-magnum removed the old fallback method of configuring
the service user with [1], and added a new one using keystone_auth
with [2]. This commit uses the new magnum::keystone::keystone_auth
class to fix the configuration.

Note that the change started in stable/ussuri, so this patch should
be backported all the way back to that release.

Closes-Bug: #1902252

[1] - https://review.opendev.org/659774
[2] - https://review.opendev.org/758981

Change-Id: I81464c3267190b3d21b99796ac42f30b6bc86cb7