Move nova host discovery to a new controller_post step
Currently, host discovery is executed from the compute nodes after each
compute setup. This will be a problem in future as it requires db access
and credentials in nova config file in the compute nodes which is not
recommended and will be forbiden in future [1].

This patch is removing it from the compute.pp template and moving it to
a new step controller_post that will be executed in the controller after
executing the rest of puppet manifests.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-November/018875.html

Change-Id: I3ed2780b55ff6944aa7bf40f259f9c4e03389373