Unpin non-openstack puppet-modules in master branch
In master branch we should be following master branches of non-openstack
puppet modules as we do in RDO. We should only pin when creating stable
branches or in justified cases when master branch breaks packstack and
can't be fixed in packstack side immediately.

I keep puppet-systemd pinned as it has introduced a backwards
incompatible change breaking puppet-rabbitmq.

Change-Id: I0cb5f07be434ed5ce417007373113d8addec0944