Add CentOS Stream 9 preview support
CentOS is publishing some preview composes of CentOS Stream 9. In RDO,
we are working to get ready for it ASAP.

This patch is proposing some changes needed for current status of
CentOS9 so that i can use this patch for it:

- Set rsync to be executed without xinetd (xinetd is not in CS9)
- Added the creation of initiatorname.iscsi using openstacklib::iscsid

It's also creating packstack-centos9 and packstack-multinode-centos9
base jobs so that we can start running packstack jobs in third-party
zuul servers.

Change-Id: I64edd1876ccc32f98ab2330024d9ab513f42a158