Configure and Start RabbitMQ before starting OpenStack services
RabbitMQ should be up and confugred before starting OpenStack services.
This change introduces ordering constraints so that all OpenStack
services are started after RabbitMQ is configured and started.

Also, a change in puppet-placement [1] has changed the default url for
placement service. This patch is also updating the keystone endpoint for
placement to align with this new default.

[1] https://review.opendev.org/c/openstack/puppet-placement/+/819052

Change-Id: I9ad973e900ccd30e41c2166515f564eed49e2a90