Replace hiera by lookup
The hiera function is deprecated and does not work with the latest
hieradata version 5. It should be replaced by the new lookup
function[1].

[1] https://puppet.com/docs/puppet/7/hiera_automatic.html

With the lookup function, we can define value type and merge behavior,
but these are kept default at this moment to limit scope of this change
to just simple replacement. Adding value type might be useful to make
sure the value is in expected type (especially when a boolean value is
expected), but we will revisit that later.

example:
lookup(<NAME>, [<VALUE TYPE>], [<MERGE BEHAVIOR>], [<DEFAULT VALUE>])

This also replaces the hiera_array function which is also deprecated,
according to the guideline[2].

[2] https://puppet.com/docs/puppet/7/hiera_migrate.html#updated_classic_hiera_function_calls

Change-Id: Ic9930107fbc68cba3432f4424f113071325efcb7