Replace deprecated defined resource type of Manila
Since [1], 'manila::service_instance' has been replaced
by 'manila::backend::service_instance'.

We have also to remove the use of
'use_guestagent_template' as it was removed in [2]

Finally, since tox4 we need to set skipdist to False
in order to install the package (needed for pbr to
get the version). Previously with tox3, it was set to
True but it was a bug apprently.

[1] https://review.opendev.org/c/openstack/puppet-manila/+/873479
[2] https://review.opendev.org/c/openstack/puppet-trove/+/874367

Change-Id: I464675a5f78b1efcc3a277da4772dca5690dd842