Enable snapshots in NFS backends and add nova section
Snapshots feature is disabled in nfs by default in cinder because of issues with old
versions of libvirt [1] which are not found in centos8 and centos9, which
are the versions supported since long time ago.

Also, we are missing the [nova] section in cinder.conf which is required
in some cinder operations, as snapshots with nfs backend.

[1] https://github.com/openstack/cinder/commit/2d77a7a87d625f47deab4955298dad9d21a93f26

Change-Id: Ie70d0f242220b513bab37befedb9ab785e08d077
(cherry picked from commit bd1e892301bfbb392c870a314dcbbb404d5a1e99)